+++ pkg/apis/deployment/v2alpha1/conditions.go
@@ -74,6 +74,8 @@ const (
 	ConditionTypeScaleDownCandidate ConditionType = "ScaleDownCandidate"
 	// ConditionTypeUpgradeFailed indicates that upgrade failed
 	ConditionTypeUpgradeFailed ConditionType = "UpgradeFailed"
+	// ConditionTypeArchitectureMismatch indicates that the member has a different architecture than the deployment.
+	ConditionTypeArchitectureMismatch ConditionType = "ArchitectureMismatch"
 
 	// ConditionTypeMemberMaintenanceMode indicates that Maintenance is enabled on particular member
 	ConditionTypeMemberMaintenanceMode ConditionType = "MemberMaintenanceMode"