+++ pkg/deployment/member/phase_updates.go
@@ -101,6 +101,7 @@ func removeMemberConditionsMapFunc(m *api.MemberStatus) {
 	m.Conditions.Remove(api.ConditionTypeTopologyAware)
 	m.Conditions.Remove(api.MemberReplacementRequired)
 	m.Conditions.Remove(api.ConditionTypePVCResizePending)
+	m.Conditions.Remove(api.ConditionTypeArchitectureMismatch)
 
 	m.RemoveTerminationsBefore(time.Now().Add(-1 * recentTerminationsKeepPeriod))